refactor: 파일 분배 가편성
